backend/requirements.txt
Django>=3.1,<3.2
djangorestframework>=3.11.1,<3.12.0
djangorestframework-simplejwt>=4.4.0,<4.5.0
django-cors-headers==3.4.0

flake8>=3.8.3,<3.8.4
